// restat/restat-frontend/src/services/types/onboarding-steps.ts
export enum IOnboardingStepType {
  addYourTeam = "addYourTeam",
  buildYourPortfolio = "buildYourPortfolio",
  createCustomTemplates = "createCustomTemplates",
  connectYourProfiles = "connectYourProfiles",
  expandYourReach = "expandYourReach",
  speedUpYourWorkflow = "speedUpYourWorkflow",
  syncYourBids = "syncYourBids",
  efficientDealsManagement = "efficientDealsManagement",
  automateYourWorkflow = "automateYourWorkflow",
  centralizeContactInformation = "centralizeContactInformation",
  streamlineYourProspecting = "streamlineYourProspecting",
  manageBusinessProfileSettings = "manageBusinessProfileSettings",
  analyzeYourPerformance = "analyzeYourPerformance",
}